+++ Dosier_LED_Ver1_3_Universalplatinen_Controller/typedef.h
@@ -14,7 +14,7 @@ typedef struct {
   boolean Dosiernachfuell;
   boolean Dosiermanuell;
   long Endtime;
-  byte Dosierspeed;
+  int Dosierspeed;
 } PUMP;
 
 typedef struct {
